Resolve syntax errors with the examples (#44)
When development work for some of the refactoring was finishing, GitHub Actions was disabled for a short period. This meant that some things that would have been caught by CI got through (format mistakes + invalid usages in examples).

This PR resolves format mistakes and ensures that the examples are up to date.