Remove workaround for terraform-11 (#45)
Remove the old version of the resources that worked for pre-Terraform 0.12. 

As terraform didn't have support for sets until 0.12, this meant that per-user resources would _shuffle_. This resulted in a notification that they were added to an app.

To avoid this issue related to no `set` support, a workaround resource was created. Now that we have `set` support, and have shifted everything over, we can remove these old resources.